Made all contact routes private
- Made `/api/contacts` routes private.
- Updated access token expiry time to 15m.
- Handle "Access token is missing" error case in validateTokenHandler.js.
- Added `user_id` field in contacts schema.
- Added middleware for access token validation.

+++ controllers/contactController.js
@@ -3,15 +3,15 @@ const Contact = require("../models/contactModel");
 
 //@description Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getAllContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 });
 
 //@description Get contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact) {
@@ -23,25 +23,34 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@description Create new contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
     const {name, email, phone} = req.body;
     if(!name || !email || !phone) {
         res.status(400);
         throw new Error("All field are mandatory.");
     }
-    const contact = await Contact.create(req.body);
+    const contact = await Contact.create({
+        name,
+        email,
+        phone,
+        user_id: req.user.id,
+    });
     res.status(201).json(contact);
 });
 
 //@description Edit complete contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const putContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact) {
         res.statusCode(404);
         throw new Error("Contact not found");
+    }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts");
     }
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
@@ -53,7 +62,7 @@ const putContact = asyncHandler(async (req, res) => {
 
 //@description Edit contact partially
 //@route PATCH /api/contacts/:id
-//@access public
+//@access private
 const patchContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact) {
@@ -70,12 +79,16 @@ const patchContact = asyncHandler(async (req, res) => {
 
 //@description Delete contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact) {
         res.statusCode(404);
         throw new Error("Contact not found");
+    }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts");
     }
     await Contact.findByIdAndDelete(req.params.id);
     res.status(200).json(contact);
